Sample programs using Python
Some python programs including a program that opens a file for input and prints the count of four-letter words in it, to create a dictionary of phone numbers and names of five persons. Display the contents of the dictionary in alphabetical order of names, to check if a given string is a palindrome or not, without reversing the original string, reads an integer N from the keyboard and then calls a user-defined function to compute and displays the sum of the numbers from N to (2*N ) if N is non-negative. if n is negative then displays the sum of the numbers from (2* N) to N. The starting and ending points are included in the sum, to raise numbers stored in a file and separate the even and odd numbers into two files named Even.txt and Odd.txt respectively. Display both the files